move preprocess and postprocess
