+++ qiskit/algorithms/gradients/base_estimator_gradient.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from collections import defaultdict
 from collections.abc import Sequence
-from copy import copy, deepcopy
+from copy import copy
 
 import numpy as np
 
@@ -181,27 +181,50 @@ class BaseEstimatorGradient(ABC):
                         f"{i}-th circuit."
                     )
 
+    def _preprocess(
+        self,
+        circuits: Sequence[QuantumCircuit],
+        parameter_values: Sequence[Sequence[float]],
+        parameters: Sequence[Sequence[Parameter] | None],
+        supported_gates: Sequence[str],
+    ):
+        """Preprocess the gradient."""
+        g_circuits, g_parameter_values, g_parameters = [], [], []
+        for circuit, parameter_value_, parameters_ in zip(circuits, parameter_values, parameters):
+            if not _circuit_key(circuit) in self._gradient_circuit_cache:
+                transpiled_circuit = transpile(
+                    circuit, basis_gates=supported_gates, optimization_level=0
+                )
+                self._gradient_circuit_cache[
+                    _circuit_key(circuit)
+                ] = self._assign_unique_parameters(transpiled_circuit)
+            gradient_circuit = self._gradient_circuit_cache[_circuit_key(circuit)]
+            g_circuits.append(gradient_circuit.gradient_circuit)
+            g_parameter_values.append(
+                self._make_gradient_parameter_values(circuit, gradient_circuit, parameter_value_)
+            )
+            g_parameters.append(
+                self._make_gradient_parameters(circuit, gradient_circuit, parameters_)
+            )
+        return g_circuits, g_parameter_values, g_parameters
+
     def _assign_unique_parameters(
         self,
         circuit: QuantumCircuit,
-        supported_gates: Sequence[str] = None,
     ) -> GradientCircuit:
         """Assign unique parameters to the circuit.
 
         Args:
             circuit: The circuit to assign unique parameters.
-            supported_gates: The list of gate names to be supported. If None, all gates are
-                supported.
 
         Returns:
             The circuit with unique parameters.
         """
-        circuit2 = transpile(circuit, basis_gates=supported_gates, optimization_level=0)
-        gradient_circuit = circuit2.copy_empty_like(f"{circuit.name}_gradient")
+        gradient_circuit = circuit.copy_empty_like(f"{circuit.name}_gradient")
         parameter_map = defaultdict(list)
         gradient_parameter_map = {}
         num_gradient_parameters = 0
-        for instruction, qargs, cargs in circuit2.data:
+        for instruction, qargs, cargs in circuit.data:
             if instruction.is_parameterized():
                 new_inst_params = []
                 for angle in instruction.params:
@@ -214,7 +237,7 @@ class BaseEstimatorGradient(ABC):
                 instruction.params = new_inst_params
             gradient_circuit.append(instruction, qargs, cargs)
         # For the global phase
-        gradient_circuit.global_phase = circuit2.global_phase
+        gradient_circuit.global_phase = circuit.global_phase
         if isinstance(gradient_circuit.global_phase, ParameterExpression):
             substitution_map = {}
             for parameter in gradient_circuit.global_phase.parameters:
@@ -229,6 +252,61 @@ class BaseEstimatorGradient(ABC):
 
         return GradientCircuit(gradient_circuit, parameter_map, gradient_parameter_map)
 
+    def _make_gradient_parameter_values(
+        self,
+        circuit: QuantumCircuit,
+        gradient_circuit: GradientCircuit,
+        parameter_values: np.ndarray,
+    ) -> np.ndarray:
+        """Makes parameter values for the gradient circuit.
+
+        Args:
+            circuit: The original quantum circuit
+            gradient_circuit: The gradient circuit
+            parameter_values: The parameter values for the original circuit
+            parameter_set: The parameter set to calculate gradients
+
+        Returns:
+            The parameter values for the gradient circuit.
+        """
+        g_circuit = gradient_circuit.gradient_circuit
+        g_parameter_values = np.zeros(len(g_circuit.parameters))
+        for i, g_parameter in enumerate(g_circuit.parameters):
+            expr = gradient_circuit.gradient_parameter_map[g_parameter]
+            bound_expr = expr.bind(
+                {p: parameter_values[circuit.parameters.data.index(p)] for p in expr.parameters}
+            )
+
+            g_parameter_values[i] = float(bound_expr)
+        return g_parameter_values
+
+    def _make_gradient_parameters(
+        self,
+        circuit: QuantumCircuit,
+        gradient_circuit: GradientCircuit,
+        parameters: Sequence[Parameter] | None,
+    ) -> Sequence[Parameter] | None:
+        """Makes parameters for the gradient circuit.
+
+        Args:
+            circuit: The original quantum circuit
+            gradient_circuit: The gradient circuit
+            parameters: The parameters for the original circuit
+
+        Returns:
+            The parameters for the gradient circuit.
+        """
+        if parameters is None:
+            return None
+
+        g_parameters = []
+        for parameter in circuit.parameters:
+            if parameter in parameters:
+                g_parameters.extend(
+                    g_parameter for g_parameter, _ in gradient_circuit.parameter_map[parameter]
+                )
+        return list(set(g_parameters))
+
     def _make_parameter_set(self, circuit: QuantumCircuit, parameters: Sequence[Parameter]):
         """Make a set of parameters from ``parameters`` that are in ``circuit``.
 
@@ -241,6 +319,52 @@ class BaseEstimatorGradient(ABC):
             is returned.
         """
         return set(circuit.parameters) if parameters is None else set(parameters)
+
+    def _postprocess(
+        self,
+        results: Sequence[EstimatorGradientResult],
+        circuits: Sequence[QuantumCircuit],
+        parameter_values: Sequence[Sequence[float]],
+        parameters: Sequence[Sequence[Parameter] | None],
+    ) -> EstimatorGradientResult:
+        """Postprocess the gradient."""
+        original_gradients, original_metadata = [], []
+        for circuit, parameter_values_, parameters_, gradient in zip(
+            circuits, parameter_values, parameters, results.gradients
+        ):
+            parameter_set = self._make_parameter_set(circuit, parameters_)
+            original_gradient = np.zeros(len(parameter_set))
+            gradient_circuit = self._gradient_circuit_cache[_circuit_key(circuit)]
+            g_parameters_ = self._make_gradient_parameters(circuit, gradient_circuit, parameters_)
+            g_parameter_set = self._make_parameter_set(
+                gradient_circuit.gradient_circuit, g_parameters_
+            )
+            result_indices_ = [param for param in circuit.parameters if param in parameter_set]
+            g_result_indices_ = [
+                param
+                for param in gradient_circuit.gradient_circuit.parameters
+                if param in g_parameter_set
+            ]
+            g_result_indices = {param: i for i, param in enumerate(g_result_indices_)}
+
+            for i, parameter in enumerate(result_indices_):
+                for g_parameter, coeff in gradient_circuit.parameter_map[parameter]:
+                    if isinstance(coeff, ParameterExpression):
+                        local_map = {
+                            p: parameter_values_[circuit.parameters.data.index(p)]
+                            for p in coeff.parameters
+                        }
+                        bound_coeff = coeff.bind(local_map)
+                    else:
+                        bound_coeff = coeff
+                    original_gradient[i] += bound_coeff * gradient[g_result_indices[g_parameter]]
+            original_gradients.append(original_gradient)
+            original_metadata.append(
+                [{"parameters": [p for p in circuit.parameters if p in parameter_set]}]
+            )
+        return EstimatorGradientResult(
+            gradients=original_gradients, metadata=results.metadata, options=results.options
+        )
 
     @property
     def options(self) -> Options:
