+++ qiskit/algorithms/gradients/param_shift_estimator_gradient.py
@@ -17,23 +17,16 @@ from __future__ import annotations
 
 from typing import Sequence
 
-import numpy as np
-
 from qiskit.algorithms import AlgorithmError
-from qiskit.circuit import Parameter, QuantumCircuit, ParameterExpression
+from qiskit.circuit import Parameter, QuantumCircuit
 from qiskit.opflow import PauliSumOp
 from qiskit.primitives import BaseEstimator
-from qiskit.primitives.utils import _circuit_key
 from qiskit.providers import Options
 from qiskit.quantum_info.operators.base_operator import BaseOperator
 
 from .base_estimator_gradient import BaseEstimatorGradient
 from .estimator_gradient_result import EstimatorGradientResult
-from .utils import (
-    _make_param_shift_parameter_values,
-    _make_gradient_parameter_values,
-    _make_gradient_parameters,
-)
+from .utils import _make_param_shift_parameter_values
 
 
 class ParamShiftEstimatorGradient(BaseEstimatorGradient):
@@ -78,81 +71,12 @@ class ParamShiftEstimatorGradient(BaseEstimatorGradient):
     ) -> EstimatorGradientResult:
         """Compute the gradients of the expectation values by the parameter shift rule."""
         g_circuits, g_parameter_values, g_parameters = self._preprocess(
-            circuits, parameter_values, parameters
+            circuits, parameter_values, parameters, self._SUPPORTED_GATES
         )
         results = self._run_unique(
             g_circuits, observables, g_parameter_values, g_parameters, **options
         )
-        return self._postprocess(results, circuits, parameter_values, parameters, g_parameters)
-
-    def _preprocess(
-        self,
-        circuits: Sequence[QuantumCircuit],
-        parameter_values: Sequence[Sequence[float]],
-        parameters: Sequence[Sequence[Parameter] | None],
-    ):
-        """Preprocess the gradient."""
-        g_circuits, g_parameter_values, g_parameters = [], [], []
-        for circuit, parameter_value_, parameters_ in zip(circuits, parameter_values, parameters):
-            if not _circuit_key(circuit) in self._gradient_circuit_cache:
-                self._gradient_circuit_cache[
-                    _circuit_key(circuit)
-                ] = self._assign_unique_parameters(circuit, supported_gates=self._SUPPORTED_GATES)
-            gradient_circuit = self._gradient_circuit_cache[_circuit_key(circuit)]
-            g_circuits.append(gradient_circuit.gradient_circuit)
-            g_parameter_values.append(
-                _make_gradient_parameter_values(circuit, gradient_circuit, parameter_value_)
-            )
-            g_parameters.append(_make_gradient_parameters(circuit, gradient_circuit, parameters_))
-        return g_circuits, g_parameter_values, g_parameters
-
-    def _postprocess(
-        self,
-        results: Sequence[EstimatorGradientResult],
-        circuits: Sequence[QuantumCircuit],
-        parameter_values: Sequence[Sequence[float]],
-        parameters: Sequence[Sequence[Parameter] | None],
-        g_parameters: Sequence[Sequence[Parameter] | None],
-    ) -> EstimatorGradientResult:
-        """Postprocess the gradient."""
-        original_gradients, original_metadata = [], []
-        for circuit, parameter_values_, parameters_, g_parameters_, gradient in zip(
-            circuits, parameter_values, parameters, g_parameters, results.gradients
-        ):
-            parameter_set = self._make_parameter_set(circuit, parameters_)
-            original_gradient = np.zeros(len(parameter_set))
-            gradient_circuit = self._gradient_circuit_cache[_circuit_key(circuit)]
-            g_parameter_set = self._make_parameter_set(
-                gradient_circuit.gradient_circuit, g_parameters_
-            )
-            result_indices_ = [param for param in circuit.parameters if param in parameter_set]
-            g_result_indices_ = [
-                param
-                for param in gradient_circuit.gradient_circuit.parameters
-                if param in g_parameter_set
-            ]
-            g_result_indices = {param: i for i, param in enumerate(g_result_indices_)}
-
-            for i, parameter in enumerate(result_indices_):
-                for g_parameter, coeff in gradient_circuit.parameter_map[parameter]:
-                    if isinstance(coeff, ParameterExpression):
-                        local_map = {
-                            p: parameter_values_[circuit.parameters.data.index(p)]
-                            for p in coeff.parameters
-                        }
-                        bound_coeff = coeff.bind(local_map)
-                    else:
-                        bound_coeff = coeff
-                    original_gradient[i] += (
-                        bound_coeff * gradient[g_result_indices[g_parameter]]
-                    )
-            original_gradients.append(original_gradient)
-            original_metadata.append(
-                [{"parameters": [p for p in circuit.parameters if p in parameter_set]}]
-            )
-        return EstimatorGradientResult(
-            gradients=original_gradients, metadata=results.metadata, options=results.options
-        )
+        return self._postprocess(results, circuits, parameter_values, parameters)
 
     def _run_unique(
         self,
